Initial work on the Optional Parenting System.
This interface will allow ARenderable instances to be linked to a
parent and have their position offset. The system is optional because I
don't want to fall into mass-container zealotry.

A document will be created to talk about the API and show specifics.

+++ engine/src/ASprite.cpp
@@ -148,7 +148,7 @@ void ASprite::AdvanceFrame()
         }
         else
         {
-            currentAnim->currentFrameIndex = currentAnim->Textures->Count() - 1;
+            currentAnim->currentFrameIndex = (int)currentAnim->Textures->Count() - 1;
         }
     }
 }
